requirements.txt
tensorflow>=2.13
keras>=2.13
numpy>=1.23
pandas>=1.5
matplotlib>=3.7
scikit-learn>=1.2
pillow>=9.5
shap>=0.44
ipywidgets>=8.0
jupyterlab>=3.6
tqdm>=4.65
opencv-python>=4.7